requirements.txt
# UI Framework
PySide6>=6.5.0

# OpenGL & 3D Math
PyOpenGL>=3.1.6
PyGLM>=2.7.0

# Matrix Operations & Data Processing
numpy>=1.24.0